Remove the need for summing of the ucode task completely from the RSP. Added workaround for pokemon stadium audio.

+++ rsp_hle/hle.h
@@ -45,12 +45,12 @@
 // types
 typedef unsigned char		u8;
 typedef unsigned short		u16;
-typedef unsigned long		u32;
+typedef unsigned int		u32;
 typedef unsigned long long	u64;
 
 typedef signed char			s8;
 typedef signed short		s16;
-typedef signed long			s32;
+typedef signed int			s32;
 typedef signed long long	s64;
 
 //#define ACMD_SIZE               32
@@ -75,29 +75,29 @@ extern RSP_INFO rsp;
 
 typedef struct
 {
-   unsigned long type;
-   unsigned long flags;
+   u32 type;
+   u32 flags;
 
-   unsigned long ucode_boot;
-   unsigned long ucode_boot_size;
+   u32 ucode_boot;
+   u32 ucode_boot_size;
 
-   unsigned long ucode;
-   unsigned long ucode_size;
+   u32 ucode;
+   u32 ucode_size;
 
-   unsigned long ucode_data;
-   unsigned long ucode_data_size;
+   u32 ucode_data;
+   u32 ucode_data_size;
 
-   unsigned long dram_stack;
-   unsigned long dram_stack_size;
+   u32 dram_stack;
+   u32 dram_stack_size;
 
-   unsigned long output_buff;
-   unsigned long output_buff_size;
+   u32 output_buff;
+   u32 output_buff_size;
 
-   unsigned long data_ptr;
-   unsigned long data_size;
+   u32 data_ptr;
+   u32 data_size;
 
-   unsigned long yield_data_ptr;
-   unsigned long yield_data_size;
+   u32 yield_data_ptr;
+   u32 yield_data_size;
 } OSTask_t;
 
 void jpg_uncompress(OSTask_t *task);
